+++ docs/changelog.md
@@ -63,6 +63,10 @@
 
 ## 更新说明详情
 
+1.10.83 add: add stock_industry_clf_hist_sw interface
+
+    1. 新增 stock_industry_clf_hist_sw 接口
+
 1.10.82 fix: fix currency_boc_sina interface
 
     1. 修复 currency_boc_sina 接口
@@ -2755,6 +2759,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.10.83 add: add stock_industry_clf_hist_sw interface
 
 1.10.82 fix: fix currency_boc_sina interface